requirements.txt
cape-privacy==0.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
numpy==1.18.1
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
urllib3==1.25.10
validators==0.15.0
